requirements.txt
Django==6.0.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-filter==25.1
Pillow==12.1.1
dj-database-url==3.0.1
gunicorn==23.0.0
psycopg[binary]==3.2.10
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
sqlparse==0.5.5
whitenoise==6.11.0
